# Return undefined for empty configurations