--- a/bind.js
+++ b/bind.js
@@ -32,7 +32,7 @@
 			}
 		}
 
-		return classes.join(' ');
+		return classes[0] && classes.join(' ');
 	}
 
 	if (typeof module !== 'undefined' && module.exports) {