--- a/tests/index.js
+++ b/tests/index.js
@@ -27,8 +27,8 @@ describe('classNames', function () {
 		assert.equal(classNames('', 'b', {}, ''), 'b');
 	});
 
-	it('returns an empty string for an empty configuration', function () {
-		assert.equal(classNames({}), '');
+	it('returns undefined for an empty configuration', function () {
+		assert.equal(classNames({}), undefined);
 	});
 
 	it('supports an array of class names', function () {