@JKillian 2016-04-06
This could be a breaking change in other cases also. Any code that assumes a string is returned would break. Why not just wrap `classNames` in your own function and use that across your codebase?