@hx 2016-04-07
Thanks @JKillian. I did wrap it, but joining an array into a string and then discarding it is slower than not joining it at all.

Consider this:

```javascript
<ul>
  {items.map(item => (
    <li key={item.id} className={classNames({active: activeItem === item})}>
      {item.text}
    </li>
  )}
</ul>
```

If `items` is an especially long list, you're calling `.join` on `length - 1` empty arrays. This makes the PR a win for terseness *and* performance.

I can't think of any breaking scenarios in which `undefined` vs `''` would be a bad thing. The worse is you end up with `class="undefined"` in your DOM (and only if you're implausibly casting the return value to a string), which is harmless unless you have an implausible CSS rule like `.undefined { font-size: 2000% }`, or an equally implausible query like `[class=""]`.

In any case, if versioning policy doesn't allow breaking changes on minor version bumps, maybe keep it aside for version 3.0.