@dcousens 2017-05-17
Closing in favour of @JKillian's proposal (DIY ` || undefined`),  relevant discussion: https://github.com/JedWatson/classnames/issues/115